+++ tomcat/src/main/java/org/apache/coyote/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.exception;
+package org.apache.coyote.exception;
 
 public class UnauthorizedException extends RuntimeException {
     private final String message;
