+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,20 +1,20 @@
 package org.apache.coyote.http11;
 
-import static org.apache.coyote.http11.request.Method.GET;
-import static org.apache.coyote.http11.request.Method.POST;
-import static org.apache.coyote.http11.request.RequestHeader.ACCEPT;
-import static org.apache.coyote.http11.response.ResponseHeader.LOCATION;
-import static org.apache.coyote.http11.response.ResponseHeader.SET_COOKIE;
-import static org.apache.coyote.http11.response.Status.FOUND;
-import static org.apache.coyote.http11.response.Status.INTERNAL_SERVER_ERROR;
-import static org.apache.coyote.http11.response.Status.NOT_FOUND;
-import static org.apache.coyote.http11.response.Status.OK;
-import static org.apache.coyote.http11.response.Status.UNAUTHORIZED;
-import static org.apache.coyote.http11.utils.Constant.BASE_PATH;
-import static org.apache.coyote.http11.utils.Constant.COOKIE_DELIMITER;
-import static org.apache.coyote.http11.utils.Constant.EMPTY;
-import static org.apache.coyote.http11.utils.Constant.LINE_SEPARATOR;
-import static org.apache.coyote.http11.utils.Parser.parseFormData;
+import static org.apache.coyote.request.Method.GET;
+import static org.apache.coyote.request.Method.POST;
+import static org.apache.coyote.request.RequestHeader.ACCEPT;
+import static org.apache.coyote.response.ResponseHeader.LOCATION;
+import static org.apache.coyote.response.ResponseHeader.SET_COOKIE;
+import static org.apache.coyote.response.Status.FOUND;
+import static org.apache.coyote.response.Status.INTERNAL_SERVER_ERROR;
+import static org.apache.coyote.response.Status.NOT_FOUND;
+import static org.apache.coyote.response.Status.OK;
+import static org.apache.coyote.response.Status.UNAUTHORIZED;
+import static org.apache.coyote.utils.Constant.BASE_PATH;
+import static org.apache.coyote.utils.Constant.COOKIE_DELIMITER;
+import static org.apache.coyote.utils.Constant.EMPTY;
+import static org.apache.coyote.utils.Constant.LINE_SEPARATOR;
+import static org.apache.coyote.utils.Parser.parseFormData;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -26,14 +26,14 @@ import java.util.Map;
 import nextstep.jwp.exception.UncheckedServletException;
 import nextstep.jwp.service.UserService;
 import org.apache.coyote.Processor;
-import org.apache.coyote.http11.exception.PageNotFoundException;
-import org.apache.coyote.http11.exception.UnauthorizedException;
-import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.Method;
-import org.apache.coyote.http11.request.RequestHeader;
-import org.apache.coyote.http11.response.HttpResponse;
-import org.apache.coyote.http11.response.ResponseHeader;
-import org.apache.coyote.http11.response.Status;
+import org.apache.coyote.exception.PageNotFoundException;
+import org.apache.coyote.exception.UnauthorizedException;
+import org.apache.coyote.request.HttpRequest;
+import org.apache.coyote.request.Method;
+import org.apache.coyote.request.RequestHeader;
+import org.apache.coyote.response.HttpResponse;
+import org.apache.coyote.response.ResponseHeader;
+import org.apache.coyote.response.Status;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
