+++ tomcat/src/main/java/org/apache/coyote/response/Status.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.response;
+package org.apache.coyote.response;
 
 public enum Status {
     OK(200, "OK"),
